YapDatabase compiles on Windows

+++ YapDatabase/Utilities/YapCache.m
@@ -332,7 +332,7 @@ static const NSUInteger YapCache_Default_CountLimit = 40;
 				leastRecentCacheItem = newItem;
 			
 			YDBLogVerbose(@"key(%@) <- new, new mostRecent [%ld of %lu]",
-			              key, CFDictionaryGetCount(cfdict), (unsigned long)countLimit);
+			              key, (long)CFDictionaryGetCount(cfdict), (unsigned long)countLimit);
 		}
 	}
 	
@@ -444,7 +444,7 @@ static const NSUInteger YapCache_Default_CountLimit = 40;
 - (NSString *)description
 {
 	NSMutableString *description = [NSMutableString string];
-	[description appendFormat:@"%@, count=%ld, keys=\n", NSStringFromClass([self class]), CFDictionaryGetCount(cfdict)];
+	[description appendFormat:@"%@, count=%ld, keys=\n", NSStringFromClass([self class]), (long)CFDictionaryGetCount(cfdict)];
 	
 	YapCacheItem *item = mostRecentCacheItem;
 	NSUInteger itemIndex = 0;
